SF: Add support for 64 bit consumer usage flag
Change consumer usage query calls to get 64 bit flags.
Make modifications to accomodate 64 bit flags.

CRs-Fixed: 2223978

+++ services/surfaceflinger/DisplayUtils.h
@@ -58,7 +58,7 @@ class DisplayUtils {
             sp<IGraphicBufferProducer> &producer, sp<IGraphicBufferProducer> bqProducer,
             sp<IGraphicBufferConsumer> bqConsumer, String8 currentStateDisplayName,
             bool currentStateIsSecure);
-    bool canAllocateHwcDisplayIdForVDS(int usage);
+    bool canAllocateHwcDisplayIdForVDS(uint64_t usage);
     bool skipColorLayer(const char* layerType);
     DisplayUtils();
   private:
